+++ CHANGELOG.md
@@ -27,12 +27,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `TextLog`: `write`, `clear`
   - `TreeNode`: `expand`, `expand_all`, `collapse`, `collapse_all`, `toggle`, `toggle_all`
   - `Tree`: `clear`, `reset`
-- Replaced some private attributes with public ones in the json tree example. https://github.com/Textualize/textual/pull/2138
+- Screens with alpha in their background color will now blend with the background. https://github.com/Textualize/textual/pull/2139
+- Added "thick" border style. https://github.com/Textualize/textual/pull/2139
 
 ### Added
 
 - Added auto_scroll attribute to TextLog https://github.com/Textualize/textual/pull/2127
 - Added scroll_end switch to TextLog.write https://github.com/Textualize/textual/pull/2127
+- Added Screen.ModalScreen which prevents App from handling bindings. https://github.com/Textualize/textual/pull/2139
 
 
 ## [0.16.0] - 2023-03-22